Move highlight parsing and sources into Relego.Core
Parsing (ClippingsParser, HighlightAggregator, models) and the source layer
(IHighlightSource, Kindle/Kobo sources, detectors, resolver) move from
Relego.Cli into Relego.Core so the server can reuse them for uploaded files.
The CLI keeps its commands and TUI unchanged.
