+++ .impeccable/hook.cache.json
@@ -1,0 +1,1 @@
+{"version":1,"sessions":{"20007de0-52b2-4230-b3cf-58f05cbd6fac":{"updatedAt":1785108492205,"files":{"C:\\Users\\Krusty\\.copilot\\session-state\\20007de0-52b2-4230-b3cf-58f05cbd6fac\\files\\prototype.html":{"editCount":62,"findings":["layout-transition:305","design-system-radius:180:5px","design-system-radius:265:2px","design-system-color:368:rgba(0,0,0,0.45)","flat-type-hierarchy:78","monotonous-spacing:0:~4px used 15/22 times (68%)","em-dash-overuse:0:11 em-dashes in body text","layout-transition:307","design-system-radius:182:5px","design-system-radius:267:2px","design-system-color:370:rgba(0,0,0,0.45)","flat-type-hierarchy:80"]}}}}}
