+++ src/Relego.Cli/Tui/BookListScreen.cs
@@ -1,4 +1,4 @@
-﻿using System.Collections.ObjectModel;
+using System.Collections.ObjectModel;
 using System.Globalization;
 using Terminal.Gui.Drawing;
 using Terminal.Gui.Drivers;
@@ -6,6 +6,7 @@ using Terminal.Gui.Input;
 using Terminal.Gui.ViewBase;
 using Terminal.Gui.Views;
 using Relego.Cli.Infrastructure;
+using Relego.Core.Sources;
 using Relego.Cli.Import;
 using Relego.Cli.Tui.ViewModels;
 using Relego.Core.Contracts;
