+++ src/Relego.Core/Relego.Core.csproj
@@ -1,10 +1,33 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
+  <Import Project="..\PackageVersions.props" />
+
   <PropertyGroup>
-    <Version>0.7.4</Version>
+    <Version>0.8.0</Version>
     <TargetFramework>net10.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Data.Sqlite" Version="$(MicrosoftDataSqliteVersion)" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="10.0.7" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <InternalsVisibleTo Include="Relego.Tests" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <!--
+      SQLitePCLRaw.lib.e_sqlite3 (transitive via Microsoft.Data.Sqlite) is flagged by
+      GHSA-2m69-gcr7-jv3q (SQLite < 3.50.2 memory corruption, CVE-2025-6965).
+      Suppressed here to match the mitigation applied in Relego.Server.csproj (SourceGear.sqlite3
+      3.50.4.5 places a patched binary that overrides the vulnerable transitive binary).
+      Without this suppression the build fails (TreatWarningsAsErrors=true).
+      Remove once SQLitePCLRaw.bundle_e_sqlite3 ships with SourceGear.sqlite3 as its native dep.
+    -->
+    <NuGetAuditSuppress Include="https://github.com/advisories/GHSA-2m69-gcr7-jv3q" />
+  </ItemGroup>
 
 </Project>
