+++ src/Relego.Tests/Sources/MultiSourceImportTests.cs
@@ -1,9 +1,9 @@
-﻿using System.Net;
+using System.Net;
 using System.Text;
 using Microsoft.Extensions.Logging.Abstractions;
 using Relego.Cli.Import;
 using Relego.Cli.Infrastructure;
-using Relego.Cli.Sources;
+using Relego.Core.Sources;
 using Relego.Tests.Sources.Support;
 
 namespace Relego.Tests.Sources;
